Modify RobotImporter within Gem
