+++ Gems/ROS2RobotImporter/Code/Include/ROS2RobotImporter/ROS2RobotImporterBus.h
@@ -1,32 +1,39 @@
-
+/*
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 #pragma once
 
-#include <ROS2RobotImporter/ROS2RobotImporterTypeIds.h>
-
 #include <AzCore/EBus/EBus.h>
-#include <AzCore/Interface/Interface.h>
+#include <AzCore/EBus/Policies.h>
+#include <AzCore/std/string/string.h>
+#include <ROS2RobotImporter/SDFormatModelPluginImporterHook.h>
+#include <ROS2RobotImporter/SDFormatSensorImporterHook.h>
 
-namespace ROS2RobotImporter
+namespace ROS2
 {
-    class ROS2RobotImporterRequests
+    class RobotImporterRequest : public AZ::EBusTraits
     {
     public:
-        AZ_RTTI(ROS2RobotImporterRequests, ROS2RobotImporterRequestsTypeId);
-        virtual ~ROS2RobotImporterRequests() = default;
-        // Put your public methods here
-    };
-
-    class ROS2RobotImporterBusTraits : public AZ::EBusTraits
-    {
-    public:
-        //////////////////////////////////////////////////////////////////////////
-        // EBusTraits overrides
         static constexpr AZ::EBusHandlerPolicy HandlerPolicy = AZ::EBusHandlerPolicy::Single;
         static constexpr AZ::EBusAddressPolicy AddressPolicy = AZ::EBusAddressPolicy::Single;
-        //////////////////////////////////////////////////////////////////////////
+
+        //! Generate prefab from the urdf file.
+        //! @param filePath The path of the urdf file
+        //! @param importAssetWithUrdf If true, the assets referenced in the urdf file will be imported
+        //! @param useArticulation If true, the prefab will be generated with articulation
+        virtual bool GeneratePrefabFromFile(const AZStd::string_view filePath, bool importAssetWithUrdf, bool useArticulation) = 0;
+
+        //! Return the reference to the list of sensor importer hooks
+        virtual const SDFormat::SensorImporterHooksStorage& GetSensorHooks() const = 0;
+
+        //! Return the reference to the list of model plugin importer hooks
+        virtual const SDFormat::ModelPluginImporterHooksStorage& GetModelPluginHooks() const = 0;
     };
 
-    using ROS2RobotImporterRequestBus = AZ::EBus<ROS2RobotImporterRequests, ROS2RobotImporterBusTraits>;
-    using ROS2RobotImporterInterface = AZ::Interface<ROS2RobotImporterRequests>;
+    using RobotImporterRequestBus = AZ::EBus<RobotImporterRequest>;
 
-} // namespace ROS2RobotImporter
+} // namespace ROS2
