+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/Pages/CheckAssetPage.cpp
@@ -7,12 +7,12 @@
  */
 
 #include "CheckAssetPage.h"
-#include "RobotImporter/Utils/SourceAssetsStorage.h"
 #include <AzCore/Math/MathStringConversions.h>
 #include <AzFramework/Asset/AssetSystemBus.h>
 #include <QHeaderView>
 #include <QPushButton>
 #include <QVBoxLayout>
+#include <RobotImporter/Utils/SourceAssetsStorage.h>
 
 namespace ROS2
 {
