--- Gems/ROS2RobotImporter/Code/Source/RobotImporter/ROS2RobotImporterSystemComponent.cpp
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) Contributors to the Open 3D Engine Project.
- * For complete copyright and license terms please see the LICENSE at the root of this distribution.
- *
- * SPDX-License-Identifier: Apache-2.0 OR MIT
- *
- */
-
-#include "ROS2RobotImporterSystemComponent.h"
-#include <AzCore/Serialization/EditContext.h>
-#include <AzCore/Serialization/EditContextConstants.inl>
-#include <AzCore/Serialization/SerializeContext.h>
-
-namespace ROS2
-{
-    void ROS2RobotImporterSystemComponent::Reflect(AZ::ReflectContext* context)
-    {
-        if (auto serializeContext = azrtti_cast<AZ::SerializeContext*>(context))
-        {
-            serializeContext->Class<ROS2RobotImporterSystemComponent, AZ::Component>()->Version(0);
-
-            if (AZ::EditContext* ec = serializeContext->GetEditContext())
-            {
-                ec->Class<ROS2RobotImporterSystemComponent>(
-                      "Robot Importer", "Use this component to import robot definition from supported formats")
-                    ->ClassElement(AZ::Edit::ClassElements::EditorData, "")
-                    ->Attribute(AZ::Edit::Attributes::AppearsInAddComponentMenu, AZ_CRC_CE("System"))
-                    ->Attribute(AZ::Edit::Attributes::Category, "ROS2")
-                    ->Attribute(AZ::Edit::Attributes::AutoExpand, true);
-            }
-        }
-    }
-
-    void ROS2RobotImporterSystemComponent::GetProvidedServices(AZ::ComponentDescriptor::DependencyArrayType& provided)
-    {
-        provided.push_back(AZ_CRC_CE("RobotImporterService"));
-    }
-
-    void ROS2RobotImporterSystemComponent::GetIncompatibleServices(AZ::ComponentDescriptor::DependencyArrayType& incompatible)
-    {
-        incompatible.push_back(AZ_CRC_CE("RobotImporterService"));
-    }
-} // namespace ROS2
