+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/URDF/RobotControlMaker.cpp
@@ -9,7 +9,7 @@
 #include "RobotControlMaker.h"
 
 #include <AzToolsFramework/Entity/EditorEntityHelpers.h>
-#include <ROS2/RobotImporter/RobotImporterBus.h>
+#include <ROS2RobotImporter/ROS2RobotImporterBus.h>
 #include <RobotImporter/Utils/RobotImporterUtils.h>
 
 namespace ROS2
