+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/URDF/SensorsMaker.h
@@ -12,7 +12,7 @@
 #include <AzCore/Component/EntityId.h>
 #include <AzCore/Outcome/Outcome.h>
 #include <AzCore/std/containers/set.h>
-#include <ROS2/RobotImporter/SDFormatSensorImporterHook.h>
+#include <ROS2RobotImporter/SDFormatSensorImporterHook.h>
 
 #include <sdf/sdf.hh>
