+++ Gems/ROS2RobotImporter/Code/ros2robotimporter_api_files.cmake
@@ -2,6 +2,6 @@
 set(FILES
     Include/ROS2RobotImporter/ROS2RobotImporterBus.h
     Include/ROS2RobotImporter/ROS2RobotImporterTypeIds.h
-    Include/RobotImporter/SDFormatModelPluginImporterHook.h
-    Include/RobotImporter/SDFormatSensorImporterHook.h
+    Include/ROS2RobotImporter/SDFormatModelPluginImporterHook.h
+    Include/ROS2RobotImporter/SDFormatSensorImporterHook.h
 )
