+++ Gems/ROS2RobotImporter/Code/ros2robotimporter_editor_private_files.cmake
@@ -22,8 +22,6 @@ set(FILES
     Source/RobotImporter/Pages/XacroParamsPage.h
     Source/RobotImporter/RobotImporterWidget.cpp
     Source/RobotImporter/RobotImporterWidget.h
-    Source/RobotImporter/ROS2RobotImporterEditorSystemComponent.cpp
-    Source/RobotImporter/ROS2RobotImporterEditorSystemComponent.h
     # temporarily disable import hooks for sensors and models for https://github.com/o3de/sig-simulation/pull/96
     # Source/RobotImporter/SDFormat/Hooks/ROS2AckermannModelHook.cpp
     # Source/RobotImporter/SDFormat/Hooks/ROS2CameraSensorHook.cpp
